1040/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-documentai
google-cloud-storage
gcsfs
fsspec
SQLAlchemy==1.4.1     
pg8000==1.19.5
google-cloud-secret-manager==1.0.0
